Extract reusable parser tests
